move theme-switcher
