+++ src/lib/components/ui/theme-switcher/theme-switcher.svelte
