+++ src/routes/theme-switcher/default.svelte
@@ -1,5 +1,5 @@
 <script lang="ts">
-	import ThemeSwitcher from '$lib/components/shared/theme-switcher.svelte';
+	import ThemeSwitcher from '$lib/components/ui/theme-switcher/theme-switcher.svelte';
 </script>
 
 <ThemeSwitcher />
